Fix retimer register dump false behaving
Retimer register dump failed dumping single register.
This is caused by declaring option strings inside if clause,
which leads to overwriting of the first option string.
Option "-a" was overwritten by "-v"
causing "-a" not sent to the register dump script,
so it always dump the common registers.
Moving the declaration of option strings outside the if clause.

Signed-off-by : leohu@nvidia.com
Fixes nvbug https://nvbugspro.nvidia.com/bug/4860975