The changes convert the faultlog and fdr EPOCHTIME to ms
Issue was obseverd in the logservices
URI the created timestamp was wrong because
the dbus was giving the epoch time in sec
and the conversion was in microsecond
changing it to microseconds.

Test
Ran the changes in PDC and checked the
faultlog entries timestamp. IT is
showing correct time now.

Fixes nvbug https://nvbugspro.nvidia.com/bug/4735019
Signed-off-by: Deepak Kamat <dkamat@nvidia.com>