Adding the changes to collect previous boot log
Requirement was to collect the previous boot log
at the satrt of the next boot and store in a
location in emmc as with the time the logs wraps
and we loose the valulable previous boot logs.
Need to add it as a manager dump part.

Fix added is the journal logs with -b -1
to collect the previous boot logs.

Test was done by collecting the manager
dump and previous boot log file was
present within the dump.

Fixes nvbug https://nvbugspro.nvidia.com/bug/4870065
Signed-off-by: Deepak Kamat <dkamat@nvidia.com>