Making the progress timer work for the Dump task
Issue it was seen that the dump progress task is
going from the 0 to 100 percent and not changing.

Fix was to change the time from micro sec to
sec which was causing the timer to overflow.

Test was done by capturing the dump one manager
and one system dump and observing the task
complete percentage should change from 0 to
non 100 value.

Fixes nvbug https://nvbugspro.nvidia.com/bug/4909055
Signed-off-by: Deepak Kamat <dkamat@nvidia.com>