Adding the changes to catch timeout for the manager dump
Issue we have seen is that the timeout for the manager
dump is the default timeout for the bmcweb which is
45 min so any task fails only after 45 min.

Fix we have given is we will cancel the dump
if the dump child process is failed.

Test was done by exhausting the /tmp
and then creating the dump which got
failed and task state was canceled.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5020664
Signed-off-by: Deepak Kamat <dkamat@nvidia.com>