Adding the watchdog for the PDC
Issue we are seeing is that sometime
the dump offload fails and the PDC
stucks after that any other dump
creation fails.

Adding a watchdog to see if the PDC
is stuck for the defined time we will
restart the service to recover.

Test Details
Test we have did is adding the sleep
in the dump offload path to simulate
the dump download stuck scenario and
let the watchdog restart the service
after the timeout happens along with
the scenario positive path scenario
where there were no issues.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5064809
Signed-off-by: Deepak Kamat <dkamat@nvidia.com>