Adding the fix for the journal collection at the skip of journal pretty
Issue we have observed is that at the journal pretty skip
we want to collect the journal normal logs but it was not
happening.

Fix we have provided is that the status for the journal command
was not handled properly so fixed it.

Test we have done by manually changing the return code for the
function to capture the error scenario.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5078587
Signed-off-by: Deepak Kamat <dkamat@nvidia.com>