Removing the dump worker log files in /tmp
Issue we have seen is that the /tmp is not
cleaned up after the nsmddbusinfo script is
ran. which causes the 14mb of the storage
loss in the /tmp

Fix we have did is cleanup happens
immediately after the cat command
that concatenates all the worker logs,
ensuring the temporary files are
removed as soon as they're no
longer needed. The -f flag ensures
the command doesn't fail if no
matching files are found.

Test is done by creating the
manager dump and checking the
/tmp after the dump creation.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5135551
Signed-off-by: Deepak Kamat's avatarDeepak Kamat <dkamat@nvidia.com>