tools: dreport: Fix MCTP Debug Data
Since the MCTP backend implementation has changed to use in-kernel MCTP,
we need to modify the dump collection plugins to suit.

Fixes JIRA https://upstream_sync

Tested:
```
Created a Manager dump on the HMC and ensured that the MCTP D-Bus tree
as well the network information is now collected in the dump file.
```

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>