Update counter dump to nsm info script
This commit renames the script responsible for collecting counter dump
information for all devices to collecting all dump information from nsmd
service. This update introduces checks to verify the existence of the
`nsmDump` and `nsmProgressCountersReader` binaries before
attempting to collect their respective data dumps. If the
binaries are not found, a warning is logged to inform the user.

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-17895
Signed-off-by: Paweł Iwaneczko <piwaneczko@nvidia.com>